third_party/WebKit/LayoutTests/fast/pagination/caret-range-outside-paged-y-rtl-vertical-rl-expected.txt

Test that hitting areas outside a paged container takes us to the right text content on some page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
To the right of page 1:
PASS document.caretRangeFromPoint(360,91).startOffset is 0
PASS document.caretRangeFromPoint(360,150).startOffset is 0
PASS document.caretRangeFromPoint(360,300).startOffset is 0
Below page 1:
PASS document.caretRangeFromPoint(400,320).startOffset is 0
PASS document.caretRangeFromPoint(330,320).startOffset is 0
PASS document.caretRangeFromPoint(290,320).startOffset is 4
PASS document.caretRangeFromPoint(250,320).startOffset is 8
Scroll to page 3
Above page 3:
PASS document.caretRangeFromPoint(330,0).startOffset is 27
PASS document.caretRangeFromPoint(290,0).startOffset is 31
PASS document.caretRangeFromPoint(250,0).startOffset is 35
xxx xxx xxx xxx xxx xxx xxx xxx xxx
